Split golang integration test into two files
